# api/openapi.yaml
openapi: 3.0.0
info:
  title: REST API Blueprint
  description: |
    A reusable, contract‑first REST API template with enterprise‑ready features.
    This specification defines the `/health` endpoint and adds authentication,
    user profile, and admin management.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /auth/verify-otp:
    post:
      summary: Verify OTP and activate account
      description: |
        For web browsers, tokens are set as httpOnly cookies. For mobile apps, tokens are also returned in the JSON response body.
      operationId: verifyOtp
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, otp]
              properties:
                email:
                  type: string
                otp:
                  type: string
                  minLength: 6
                  maxLength: 6
      responses:
        "200":
          description: Account activated – access_token and refresh_token returned in JSON and also set as httpOnly cookies.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid or expired OTP
  /auth/refresh:
    post:
      summary: Refresh access token
      description: |
        For web browsers, uses the httpOnly refresh token cookie. For mobile apps, expects refresh_token in the request body (JSON).
        Returns new tokens in JSON body and also sets new httpOnly cookies (for web).
      operationId: refreshToken
      tags:
        - Auth
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: Required for mobile apps; for web, the cookie is used.
      responses:
        "200":
          description: New access token and refresh token issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          description: Missing or invalid refresh token
...
  /auth/session:
    get:
      summary: Get current session user
      operationId: getSession
      tags:
        - Auth
      security:
        - CookieAuth: []
      responses:
        "200":
          description: Current user session
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized (missing or invalid token)
  /auth/password-reset/request:
    post:
      summary: Request password reset email
      operationId: requestPasswordReset
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
      responses:
        "202":
          description: Reset email sent (if account exists)
        "400":
          description: Invalid email
  /auth/password-reset/confirm:
    post:
      summary: Confirm password reset with token
      operationId: confirmPasswordReset
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Password updated
        "401":
          description: Invalid or expired token

  # ==================== USERS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Use this for mobile apps or direct API access. For web browsers, prefer CookieAuth.
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: httpOnly cookie containing the JWT access token
  schemas:
...
    LoginResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - expires_in
        - token_type
        - user
      properties:
        access_token:
          type: string
          description: JWT access token (for mobile apps)
        refresh_token:
          type: string
          description: Refresh token (for mobile apps)
        expires_in:
          type: integer
          description: Access token validity in seconds
          example: 3600
        token_type:
          type: string
          example: Bearer
        user:
          $ref: "#/components/schemas/UserProfile"
    RefreshResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - expires_in
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
    UserProfile:
      type: object
      required:
        - id
        - username
        - email
        - role
      properties:
        id:
          type: string
        username:
          type: string
          description: Unique username
        email:
          type: string
        firstName:
          type: string
          description: User's first name (optional)
        lastName:
          type: string
          description: User's last name (optional)
        emailVerified:
          type: boolean
          description: Whether the email address has been verified
          default: false
        avatar:
          type: string
          description: URL to profile picture (optional)
        role:
          type: string
          enum: [user, admin]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...